+++ src/domain/health-check/app-health.interface.ts
@@ -3,5 +3,6 @@ import { AppStatus } from '../../application/enums/appStatus'
 export interface AppHealth {
   uptime: number
   status: AppStatus
+
   timestamp: number
 }