+++ src/domain/user/__test__/user-controller.test.ts
@@ -25,6 +25,7 @@ describe('user', () => {
   })
 
   it('returns a 400 if the user already exists', async () => {
+
     await createUser('Tu')
 
     await request(app)